refactor: internal/centralizedmetrics
