+++ internal/centralizedmetrics/common/structs.go
