+++ mobile/status.go
@@ -21,12 +21,12 @@ import (
 
 	accscommon "github.com/status-im/status-go/accounts-management/common"
 	"github.com/status-im/status-go/accounts-management/keystore"
-	"github.com/status-im/status-go/centralizedmetrics"
-	"github.com/status-im/status-go/centralizedmetrics/providers"
 	gocommon "github.com/status-im/status-go/common"
 	"github.com/status-im/status-go/crypto"
 	"github.com/status-im/status-go/crypto/types"
 	abi_spec "github.com/status-im/status-go/internal/abi-spec"
+	"github.com/status-im/status-go/internal/centralizedmetrics"
+	"github.com/status-im/status-go/internal/centralizedmetrics/providers"
 	"github.com/status-im/status-go/internal/db/multiaccounts"
 	"github.com/status-im/status-go/internal/db/multiaccounts/settings"
 	"github.com/status-im/status-go/internal/images"
