+++ protocol/requests/add_centralized_metric_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/status-im/status-go/centralizedmetrics/common"
+	"github.com/status-im/status-go/internal/centralizedmetrics/common"
 )
 
 func TestValidateAddCentralizedMetrics(t *testing.T) {
